+++ angular/src/app/account/send-sidechain/confirm/send-confirm.component.ts
@@ -1,64 +1,149 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
+import { UnspentOutputService } from 'src/app/services/unspent-output.service';
+import { Account, AccountUnspentTransactionOutput } from 'src/shared';
 import { SendService, SendSidechainService, WalletManager } from '../../../services';
+import { payments, Psbt } from '@blockcore/blockcore-js';
+import Big from 'big.js';
 
 @Component({
-    selector: 'app-account-send-confirm',
-    templateUrl: './send-confirm.component.html',
-    styleUrls: ['./send-confirm.component.css']
+  selector: 'app-account-send-confirm',
+  templateUrl: './send-confirm.component.html',
+  styleUrls: ['./send-confirm.component.css'],
 })
 export class AccountSendSidechainConfirmComponent implements OnInit, OnDestroy {
-    sub: any;
-    transaction: any;
-    error: string;
-    detailsOpen = false;
-    invalidFeeAmount = false;
-    loading = true;
-    valid = false;
+  sub: any;
+  transaction: any;
+  error: string;
+  detailsOpen = false;
+  loading = true;
+  valid = false;
 
-    constructor(
-        public sendService: SendService,
-        public sendSidechainService: SendSidechainService,
-        public walletManager: WalletManager) {
+  constructor(public sendService: SendService, public sendSidechainService: SendSidechainService, public walletManager: WalletManager, private unspentService: UnspentOutputService) {}
+
+  ngOnDestroy() {}
+
+  toggleDetails() {
+    this.detailsOpen = !this.detailsOpen;
+  }
+
+  async generateTransaction(account: Account, data: any) {
+    const targets: any[] = [
+      {
+        address: this.sendService.address,
+        value: this.sendService.amountAsSatoshi.toNumber(),
+      },
+    ];
+
+    // Add OP_RETURN output before we calculate fee size:
+    if (data != null && data != '') {
+      var buffer = Buffer.from(data);
+      const dataScript = payments.embed({ data: [buffer] });
+
+      targets.push({ script: dataScript.output, value: 0 }); // OP_RETURN always with 0 value unless you want to burn coins
+      // tx.addOutput({ script: dataScript.output, value: 0 }); // OP_RETURN always with 0 value unless you want to burn coins
     }
 
-    ngOnDestroy() {
+    let utxos: any[];
+    let unspent: AccountUnspentTransactionOutput[];
 
+    if (account.mode === 'normal') {
+      unspent = this.sendService.accountHistory.unspent;
+    } else {
+      // When performing send using a "quick" mode account, we will retrieve the UTXOs on-demand.
+
+      // Add an additional amount to ensure we get enough UTXO value to pay for fee. We don't really know at this time
+      // what the fee will actually be, so allow the UI/user to increase it if needed.
+      const extraFee = this.sendService.fee * 1000; // Add an additional 10000 sats by default, or higher if user changes fee.
+
+      const result = await this.unspentService.getUnspentByAmount(this.sendService.amountAsSatoshi.add(extraFee), account);
+      unspent = result.utxo;
+
+      // aggregatedAmount = result.amount;
+      // inputs.push(...result.utxo);
     }
 
-    toggleDetails() {
-        this.detailsOpen = !this.detailsOpen;
-    }
+    utxos = await Promise.all(
+      unspent.map(async (t): Promise<any> => {
+        const container = {} as any;
 
-    async ngOnInit() {
-        try {
-            const tx = await this.walletManager.createTransaction(
-                this.walletManager.activeWallet,
-                this.walletManager.activeAccount,
-                this.sendService.address,
-                this.sendService.changeAddress,
-                this.sendService.amountAsSatoshi,
-                this.sendService.feeAsSatoshi,
-                this.sendService.accountHistory.unspent,
-                this.sendSidechainService.sidechainAddress ?? null);
+        container.txId = t.transactionHash;
+        container.vout = t.index;
+        container.value = t.balance;
+        container.address = t.address;
 
-            this.transaction = tx;
-            this.sendService.transactionHex = tx.transactionHex;
-            this.sendService.addresses = tx.addresses;
-            this.invalidFeeAmount = this.transaction.feeRate < this.sendService.feeRate;
+        let hex = t.hex;
 
-            // TODO: Add aditional conditions here if needed for validating the transaction before allowing
-            // it to be broadcasted. Perhaps checking the network status?
-            if (this.invalidFeeAmount) {
-                this.valid = false;
-            } else {
-                this.valid = true;
-            }
-
-        } catch (err: any) {
-            console.error(err);
-            this.error = err.message;
+        // If we don't have the hex, retrieve it to be used in the transaction.
+        // This was needed when hex retrieval was removed to optimize extremely large wallets.
+        if (!hex) {
+          hex = await this.walletManager.getTransactionHex(account, t.transactionHash);
         }
 
-        this.loading = false;
+        container.nonWitnessUtxo = Buffer.from(hex, 'hex');
+        // container.witnessUtxo = {
+        //   script: Buffer.from(hex, 'hex'),
+        //   value: container.value,
+        // };
+
+        // TODO: Do we need nonWitnessUtxo and witnessUtxo?
+
+        return container;
+      })
+    );
+
+    const selectionData = await this.walletManager.selectUtxos(utxos, targets, this.sendService.fee);
+
+    // Set the selected data on the send service as we want to mark the UTXOs as spent after broadcast.
+    this.sendService.selectedData = selectionData;
+
+    let tx = await this.walletManager.createTransaction(
+      this.walletManager.activeWallet,
+      this.walletManager.activeAccount,
+      this.sendService.address,
+      this.sendService.changeAddress,
+      this.sendService.amountAsSatoshi,
+      selectionData.fee,
+      selectionData.inputs,
+      selectionData.outputs,
+      data
+    );
+
+    return tx;
+  }
+
+  async ngOnInit() {
+    try {
+      // let data = this.sendService.payment?.options.data;
+
+      // if (data != null) {
+      //   var enc = new TextEncoder(); // always utf-8
+      //   var arr = enc.encode(data);
+
+      //   // Slice the data and only read 80 bytes:
+      //   const sliced = arr.slice(0, 80); // Enforce the max length of 80 on OP_RETURN data.
+
+      //   // Decode the array back into UTF-8:
+      //   var dec = new TextDecoder('utf-8');
+      //   data = dec.decode(sliced);
+      // }
+
+      const tx = await this.generateTransaction(this.walletManager.activeAccount, this.sendSidechainService.sidechainAddress ?? null);
+
+      // Calculate and set the total, take it from the outputs wince the generat transaction
+      // can potentially have changed the output value (if sending max for example).
+      this.sendService.actualAmountAsSatoshi = Big(tx.transaction.txOutputs[0].value);
+      this.sendService.total = this.sendService.actualAmountAsSatoshi.add(tx.fee);
+
+      this.transaction = tx;
+      this.sendService.transactionHex = tx.transactionHex;
+      this.sendService.addresses = tx.addresses;
+
+      this.valid = true;
+    } catch (err: any) {
+      console.error(err);
+      this.error = err.message;
     }
+
+    this.loading = false;
+  }
 }