+++ angular/src/shared/handlers/did-request-handler.ts
@@ -1,12 +1,13 @@
 import { BackgroundManager } from '../background-manager';
-import { ActionMessage, ActionPrepareResult, ActionRequest, ActionResponse, Actions, Permission } from '../interfaces';
 import { ActionHandler, ActionState } from './action-handler';
+import { ActionMessage, ActionPrepareResult, ActionRequest, ActionResponse, Actions, DIDRequestResponse, Permission } from '../interfaces';
 import * as bitcoinMessage from 'bitcoinjs-message';
 import { HDKey } from '@scure/bip32';
 import { Network } from '../networks';
+import { BlockcoreIdentity, BlockcoreIdentityTools } from '../identity';
 
-export class SignMessageHandler implements ActionHandler {
-  action = ['signMessage'];
+export class DidRequestHandler implements ActionHandler {
+  action = ['did.request'];
 
   constructor(private backgroundManager: BackgroundManager) {}
 
@@ -17,13 +18,9 @@ export class SignMessageHandler implements ActionHandler {
   }
 
   async prepare(state: ActionState): Promise<ActionPrepareResult> {
-    if (!state.message || !state.message.request || !state.message.request.params || !state.message.request.params[0] || !state.message.request.params[0].message) {
-      throw Error('The params must include a single entry that has a message field.');
-    }
-
     return {
-      content: state.message.request.params[0].message,
-      consent: true
+      content: state.message.request.params[0].challenge,
+      consent: true,
     };
   }
 
@@ -34,13 +31,32 @@ export class SignMessageHandler implements ActionHandler {
     if (state.content) {
       let contentText = state.content;
 
-      if (typeof state.content !== 'string') {
-        contentText = JSON.stringify(state.content);
-      }
+      const proofContent = {
+        challenge: state.content,
+        origin: state.message.app,
+        type: 'web5.did.get',
+      };
 
-      let signedData = await this.signData(network, node, contentText as string);
+      const tools = new BlockcoreIdentityTools();
+      const privateKey = node.privateKey;
+      const verificationMethod = tools.getVerificationMethod(privateKey, 0, network.symbol);
+      const identity = new BlockcoreIdentity(verificationMethod);
 
-      return { key: permission.key, signature: signedData, request: state.message.request, response: state.content, content: state.content, network: network.id };
+      // "jws" or "jwt"?
+      const jws = await identity.jwt({ privateKey: privateKey, payload: proofContent });
+
+      let returnData: ActionResponse = {
+        key: permission.key,
+        request: state.message.request,
+        content: state.content,
+        network: network.id,
+        response: {
+          did: permission.key,
+          proof: jws,
+        } as DIDRequestResponse,
+      };
+
+      return returnData;
     } else {
       return { key: '', signature: '', response: null, content: null, request: state.message.request, network: network.id };
     }